Added OnTextChangedListener to CodeView
In this change, we have introduced a new listener named OnTextChangedListener in the CodeView class. This listener provides a callback method onTextChanged(String text) which gets called every time the text in the CodeView changes.

This will enable any component using CodeView to listen for changes in the text and execute any necessary actions upon text change. This could be useful for situations like live syntax checking, updating a preview, or enabling and disabling UI elements based on the content in the CodeView